refactor: moved enum logic to the use case that uses it
