--- react-client/src/types/PetDocumentRecordTypeEnum.ts
@@ -1,6 +1,0 @@
-export enum PetDocumentRecordType {
-  MedicalRecord = 1,
-  Vaccine = 2,
-  Test = 3,
-  Other = 1024,
-}
